Modified plural section handlers
