+++ packages/api/openapi/licence.yaml
@@ -182,7 +182,6 @@ paths:
           required: false
           schema:
             $ref: '#/components/schemas/roles'
-
       responses:
         200:
           description: Success
@@ -205,18 +204,24 @@ paths:
               schema:
                 $ref: '#/components/schemas/error'
 
-  '/user/{userId}/applications/applicant':
+  '/applications/applicant':
     get:
       description: Get the applicants for all applications for a user
       summary: Get the applicants for all applications for a user
-      operationId: getApplicantsByUserId
+      operationId: getApplicants
       parameters:
         - name: userId
-          in: path
+          in: query
           description: ID of the user
-          required: true
+          required: false
           schema:
             $ref: '#/components/schemas/uuid'
+        - name: role
+          in: query
+          description: ID of the user
+          required: false
+          schema:
+            $ref: '#/components/schemas/roles'
       responses:
         200:
           description: Success
@@ -239,18 +244,24 @@ paths:
               schema:
                 $ref: '#/components/schemas/error'
 
-  '/user/{userId}/applications/ecologist':
+  '/applications/ecologist':
     get:
       description: Get the ecologists for all applications for a user
       summary: Get the ecologists for all applications for a user
-      operationId: getEcologistsByUserId
+      operationId: getEcologists
       parameters:
         - name: userId
-          in: path
+          in: query
           description: ID of the user
-          required: true
+          required: false
           schema:
             $ref: '#/components/schemas/uuid'
+        - name: role
+          in: query
+          description: ID of the user
+          required: false
+          schema:
+            $ref: '#/components/schemas/roles'
       responses:
         200:
           description: Success
@@ -273,18 +284,24 @@ paths:
               schema:
                 $ref: '#/components/schemas/error'
 
-  '/user/{userId}/applications/applicant-organization':
+  '/applications/applicant-organization':
     get:
       description: Get the applicant organizations for all applications for a user
       summary: Get the applicants organizations for all applications for a user
-      operationId: getApplicantOrganizationsByUserId
+      operationId: getApplicantOrganizations
       parameters:
         - name: userId
-          in: path
+          in: query
           description: ID of the user
-          required: true
+          required: false
           schema:
             $ref: '#/components/schemas/uuid'
+        - name: role
+          in: query
+          description: ID of the user
+          required: false
+          schema:
+            $ref: '#/components/schemas/roles'
       responses:
         200:
           description: Success
@@ -307,18 +324,24 @@ paths:
               schema:
                 $ref: '#/components/schemas/error'
 
-  '/user/{userId}/applications/ecologist-organization':
+  '/applications/ecologist-organization':
     get:
       description: Get the applicant ecologist organizations for all ecologists for a user
       summary: Get the applicants ecologist organizations for all ecologists for a user
-      operationId: getEcologistOrganizationsByUserId
+      operationId: getEcologistOrganizations
       parameters:
         - name: userId
-          in: path
+          in: query
           description: ID of the user
-          required: true
+          required: false
           schema:
             $ref: '#/components/schemas/uuid'
+        - name: role
+          in: query
+          description: ID of the user
+          required: false
+          schema:
+            $ref: '#/components/schemas/roles'
       responses:
         200:
           description: Success
