+++ packages/api/src/handlers/application/application-section/get-sections.js
@@ -1,34 +1,29 @@
 import { models } from '@defra/wls-database-model'
 import { APPLICATION_JSON } from '../../../constants.js'
-import { REDIS } from '@defra/wls-connectors-lib'
-const { cache } = REDIS
 
-export const getSectionsByUserIdHandler = (section, keyFunc) => async (context, req, h) => {
+export const getSectionsHandler = (section, keyFunc) => async (context, req, h) => {
   try {
-    const { userId } = context.request.params
-    const saved = await cache.restore(req.path)
-
-    if (saved) {
-      return h.response(JSON.parse(saved))
-        .type(APPLICATION_JSON)
-        .code(200)
-    }
-
-    const applications = await models.applications.findAll({
-      where: { userId }
-    })
+    const where = req.query
+    const applications = await models.applications.findAll(Object.keys(where).length
+      ? {
+          include: {
+            model: models.applicationUsers,
+            attributes: [],
+            where
+          }
+        }
+      : {})
 
     const result = applications
       .map(a => ({ data: a.dataValues.application[section], keys: a.dataValues.targetKeys }))
       .filter(a => a.data)
       .map(a => keyFunc ? Object.assign(a.data, keyFunc(a.keys)) : a.data)
 
-    // Check the user/applicant exists
+    // Check the applicant exists
     if (!result.length) {
       return h.response().code(404)
     }
 
-    await cache.save(req.path, result)
     return h.response(result)
       .type(APPLICATION_JSON)
       .code(200)
