+++ packages/api/src/handlers/application/ecologist-organization/ecologist-organization.js
@@ -1,7 +1,7 @@
 import { getSectionHandler } from '../application-section/get-section.js'
 import { putSectionHandler } from '../application-section/put-section.js'
 import { deleteSectionHandler } from '../application-section/delete-section.js'
-import { getSectionsByUserIdHandler } from '../application-section/get-sections-by-user-id.js'
+import { getSectionsHandler } from '../application-section/get-sections.js'
 import { keyFunc, removeKeyFunc, sddsGetKeyFunc, removeSddsKeyFunc } from '../application-section/section-keys-func.js'
 
 const apiBasePath = 'ecologistOrganization'
@@ -9,8 +9,8 @@ const sddsKey = 'sddsAccountId'
 export const getApplicationEcologistOrganization =
   getSectionHandler(apiBasePath, keyFunc(apiBasePath, sddsKey))
 
-export const getEcologistOrganizationsByUserId =
-  getSectionsByUserIdHandler(apiBasePath, keyFunc(apiBasePath, sddsKey))
+export const getEcologistOrganizations =
+  getSectionsHandler(apiBasePath, keyFunc(apiBasePath, sddsKey))
 
 export const putApplicationEcologistOrganization =
   putSectionHandler(apiBasePath, sddsGetKeyFunc(sddsKey), removeSddsKeyFunc(sddsKey),
